cleaning repo and updating requirements
