--- common/hough_line_transforms.py
@@ -1,22 +1,0 @@
-import cv2
-import numpy as np
-
-img = cv2.imread('/media/matt/New Volume/missourian_pics/2017/Fall/Dump/Compton, Nate/20170918_nuervabullyingprevention_nc/20170918_nurevabullyingprevention_nc_002.JPG')
-gray = cv2.cvtColor(img,cv2.COLOR_BGR2GRAY)
-edges = cv2.Canny(gray,50,150,apertureSize = 3)
-
-lines = cv2.HoughLines(edges,1,np.pi/180,200)
-for rho,theta in lines[0]:
-    a = np.cos(theta)
-    b = np.sin(theta)
-    x0 = a*rho
-    y0 = b*rho
-    x1 = int(x0 + 1000*(-b))
-    y1 = int(y0 + 1000*(a))
-    x2 = int(x0 - 1000*(-b))
-    y2 = int(y0 - 1000*(a))
-
-    cv2.line(img,(x1,y1),(x2,y2),(0,0,255),2)
-
-cv2.imwrite('edges1.jpg', edges)
-cv2.imwrite('houghlines3.jpg',img)
