+++ requirements.txt
@@ -1,75 +1,93 @@
+anyio==3.0.1
 argon2-cffi==20.1.0
 async-generator==1.10
 attrs==20.2.0
+Babel==2.9.1
 backcall==0.2.0
-bleach==3.2.1
+bleach==3.3.0
 certifi==2020.6.20
 cffi==1.14.3
+chardet==4.0.0
+click==8.0.0
 cycler==0.10.0
 decorator==4.4.2
 defusedxml==0.6.0
 entrypoints==0.3
+Flask==2.0.0
 future==0.18.2
+idna==2.10
 imageio==2.9.0
 ipykernel==5.3.4
 ipython==7.18.1
 ipython-genutils==0.2.0
 ipywidgets==7.5.1
+itsdangerous==2.0.0
 jedi==0.17.2
-Jinja2==2.11.2
+Jinja2==3.0.0
 joblib==0.17.0
+json5==0.9.5
 jsonschema==3.2.0
-jupyter==1.0.0
-jupyter-client==6.1.7
-jupyter-console==6.2.0
-jupyter-core==4.6.3
+jupyter-client==6.1.12
+jupyter-core==4.7.1
+jupyter-server==1.7.0
+jupyterlab==3.0.15
 jupyterlab-pygments==0.1.2
+jupyterlab-server==2.5.1
 kiwisolver==1.2.0
-MarkupSafe==1.1.1
+MarkupSafe==2.0.0
 matplotlib==3.3.2
 mistune==0.8.4
+nbclassic==0.2.8
 nbclient==0.5.1
 nbconvert==6.0.7
-nbformat==5.0.8
+nbformat==5.1.3
 nest-asyncio==1.4.1
 networkx==2.5
-notebook==6.1.4
+notebook==6.3.0
 numpy==1.19.2
+opencv-python==4.4.0.46
 packaging==20.4
 pandas==1.1.3
 pandocfilters==1.4.2
 parso==0.7.1
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==8.0.0
+Pillow==8.2.0
 prometheus-client==0.8.0
 prompt-toolkit==3.0.8
+psycopg2-binary==2.8.6
 ptyprocess==0.6.0
 pycparser==2.20
-Pygments==2.7.1
+Pygments==2.9.0
 pyparsing==2.4.7
 pyrsistent==0.17.3
 python-dateutil==2.8.1
 pytz==2020.1
 PyWavelets==1.1.1
 pyzmq==19.0.2
-qtconsole==4.7.7
 QtPy==1.9.0
+requests==2.25.1
 scikit-image==0.17.2
 scikit-learn==0.23.2
 scipy==1.5.3
 Send2Trash==1.5.0
 six==1.15.0
 sklearn==0.0
+sniffio==1.2.0
 SQLAlchemy==1.3.20
+tdqm==0.0.1
 terminado==0.9.1
 testpath==0.4.4
 threadpoolctl==2.1.0
 tifffile==2020.10.1
 torch==1.6.0
 torchvision==0.7.0
-tornado==6.0.4
+tornado==6.1
+tqdm==4.59.0
 traitlets==5.0.5
+urllib3==1.26.4
 wcwidth==0.2.5
 webencodings==0.5.1
+websocket-client==0.59.0
+Werkzeug==2.0.0
 widgetsnbextension==3.5.1
